requirements.txt
transformers==4.39.3
transformers[sentencepiece]==4.39.3
datasets==2.19.0
sacrebleu==2.4.2
rouge_score==0.1.2
py7zr==0.20.8
pandas==2.2.2
nltk==3.8.1
tqdm==4.66.4
PyYAML==6.0.1
matplotlib==3.8.4
torch==2.2.2+cu121
notebook==7.2.0
boto3==1.34.113
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.95.2
uvicorn==0.22.0
Jinja2==3.1.2
-e .

#gshahText